% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{SeqtometryGUI}
\alias{SeqtometryGUI}
\title{Seqtometry GUI}
\value{
R6 object with all fields initialized
}
\description{
An R6 object encapsulating all data for passing and modification (by reference) across Shiny app components
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SO}}{A Seurat object containing a gene expression matrix that has passed QC}

\item{\code{GS}}{A named list of gene signatures}

\item{\code{IA}}{A string denoting the name of the imputed assay}

\item{\code{DT}}{Root node of data.tree for all data.
Each node contains a data.table with signature scores, gates, as well as Seurat object metadata and embeddings}

\item{\code{AP}}{Node that represents the current level (subtree) of the gating hierarchy being inspected}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqtometryGUI-new}{\code{SeqtometryGUI$new()}}
\item \href{#method-SeqtometryGUI-run_gui}{\code{SeqtometryGUI$run_gui()}}
\item \href{#method-SeqtometryGUI-clone}{\code{SeqtometryGUI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqtometryGUI-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqtometryGUI-new}{}}}
\subsection{Method \code{new()}}{
Initialize fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqtometryGUI$new(
  seurat_object,
  gene_signatures,
  assay = "RNA",
  impute = TRUE,
  reduction = NULL,
  approximate = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seurat_object}}{Seurat object: contains the data to be analyzed}

\item{\code{gene_signatures}}{list of character: gene signatures to be used for scoring}

\item{\code{assay}}{string: the name of the assay in the Seurat object to be used for analysis}

\item{\code{impute}}{boolean: whether to perform MAGIC imputation on parameter assay}

\item{\code{reduction}}{string: the name of the reduction in the Seurat object to be used for fitting the diffusion operator}

\item{\code{approximate}}{boolean: whether to perform approximate MAGIC imputation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqtometryGUI-run_gui"></a>}}
\if{latex}{\out{\hypertarget{method-SeqtometryGUI-run_gui}{}}}
\subsection{Method \code{run_gui()}}{
run a Shiny app using the data within this object's fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqtometryGUI$run_gui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqtometryGUI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqtometryGUI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqtometryGUI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
